Refactor index.js
